% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtExp.R
\name{is_log_trans}
\alias{is_log_trans}
\alias{is_log_trans<-}
\alias{is_log_trans.default}
\alias{is_log_trans<-.default}
\title{Is a dataset log-transformed?}
\usage{
is_log_trans(object)

is_log_trans(object) <- value

\method{is_log_trans}{default}(object)

\method{is_log_trans}{default}(object) <- value
}
\arguments{
\item{object}{An object with data that may have been log-transformed.}
}
\value{
\code{TRUE} if the data has been log-transformed
and \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if the data has been log-transformed
and \code{FALSE} otherwise.
}
\details{
This is an S3 generic function: methods can be defined for
it by writing functions with the naming convention
\code{is_log_trans.classname}. Functions should return
\code{TRUE} or \code{FALSE}. The default method gets and sets
the \code{is_log_trans} attribute.
}
\examples{
x <- structure("test", is_log_trans = TRUE)
is_log_trans(x)
is_log_trans(x) <- FALSE
is_log_trans(x)

}
