% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtExp.R
\name{normalize}
\alias{normalize}
\alias{normalize.ProtExp}
\title{Normalize a dataset}
\usage{
normalize(object, ...)

\method{normalize}{ProtExp}(object, ..., by)
}
\arguments{
\item{object}{An object with data to be normalized.}
}
\value{
Another object of the same class with the normalized data.
}
\description{
Performs normalization on an experimental dataset.
}
\details{
This is an S3 generic function: methods can be defined for
it by writing functions with the naming convention
\code{normalize.classname}. Functions should be appropriate
for the experiment type.

\code{normalize.ProtExp} performs normalization on a \code{ProtExp}
proteomics experiment dataset object using the median-of-medians method.
It also performs log2-transformation on the intensities if the data
has not already been log2-transformed.
}
