% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtExp.R
\name{ProtExp}
\alias{ProtExp}
\title{Proteomics Experiment}
\usage{
ProtExp(protein, feature, run, intensity, label, ..., is_log_trans = FALSE)
}
\arguments{
\item{protein}{A \code{character} vector of protein names.}

\item{feature}{A \code{character} vector of MS features.}

\item{run}{A \code{character} or \code{numeric} vector of MS runs.}

\item{intensity}{A \code{numeric} vector of MS intensities.}

\item{label}{A \code{character} or \code{factor} of MS labels.}

\item{...}{Additional columns.}

\item{is_log_trans}{\code{TRUE} or \code{FALSE} indicating whether
the intensities has been log2-transformed or not.}
}
\value{
object A \code{ProtExp} object.
}
\description{
Create a \code{ProtExp} object.
}
\details{
\code{ProtExp} is a simple S3 class inheriting from \code{data.frame}
for storing proteomics experiments. It expects columns named
`protein', `feature', `run', `intensity', and `label'. Additional
columns are also allowed. A \code{ProtExp} object also has a
\code{is_log_exp} attribute for tracking whether the intensities
have been log2-transformed or not.
}
